Fixes for OOP2

+++ mai_oop/Lab02/Hexagon.h
@@ -8,7 +8,7 @@ class Hexagon : public Figure{
 public:
     Hexagon();
     Hexagon(std::istream &is);
-    Hexagon(size_t a);
+    Hexagon(int a);
     Hexagon(const Hexagon& orig);
 
     friend bool operator==(const Hexagon& left,const Hexagon& right);
@@ -17,7 +17,7 @@ public:
 
     virtual ~Hexagon();
 private:
-    size_t side_a;
+    int side_a;
 };
 
 #endif	/* HEXAGON_H */
