+++ mai_oop/Lab02/Pentagon.h
@@ -8,7 +8,7 @@ class Pentagon : public Figure{
 public:
     Pentagon();
     Pentagon(std::istream &is);
-    Pentagon(size_t a);
+    Pentagon(int a);
     Pentagon(const Pentagon& orig);
 
     friend bool operator==(const Pentagon& left,const Pentagon& right);
@@ -17,7 +17,7 @@ public:
 
     virtual ~Pentagon();
 private:
-    size_t side_a;
+    int side_a;
 };
 
 #endif	/* PENTAGON_H */
