+++ mai_oop/Lab02/Rhombus.h
@@ -8,7 +8,7 @@ class Rhombus : public Figure{
 public:
     Rhombus();
     Rhombus(std::istream &is);
-    Rhombus(size_t d1, size_t d2);
+    Rhombus(int d1, int d2);
     Rhombus(const Rhombus& orig);
     friend std::ostream& operator<<(std::ostream& os,const Rhombus& rhombus);
     friend bool operator==(const Rhombus& left,const Rhombus& right);
@@ -17,8 +17,8 @@ public:
 
     virtual ~Rhombus();
 private:
-    size_t side_d1;
-    size_t side_d2;
+    int side_d1;
+    int side_d2;
 };
 
 #endif	/* RHOMBUS_H */
